ml/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31
openai>=1.30
sentence-transformers>=2.7
numpy>=1.26
opencv-python-headless>=4.9
gunicorn>=21.2
